Complete impelemntation to pass test

+++ src/com/company/Beverage.java
@@ -3,5 +3,5 @@ package com.company;
 public interface Beverage {
     //public String description = null;
     String getDescription();
-    int cost();
+    double cost();
 }
